refactor DAG related data existence check to the db table classes.
